AMP-5910
git-svn-id: file:///svn/trunk/amp@9412 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/dataExchange/utils/ImportBuilder.java
@@ -398,7 +398,7 @@ public class ImportBuilder {
 				
 				AmpActivityProgram ampActivityProgram = new AmpActivityProgram();
 				ampActivityProgram.setActivity(activity);
-				ampActivityProgram.setProgramPercentage((new Float(idmlProgram.getPercentage())).longValue());
+				ampActivityProgram.setProgramPercentage((new Float(idmlProgram.getPercentage())));
 				ampActivityProgram.setProgram(ampTheme);
 				
 				//ampActivityProgram.setProgramSetting()
